Created custom Include method
It's come to my attention that program_path and current_path are not reliable variables when a program is executed remotely. Unfortunately this means the only other option is to search the entire system for built in libraries. I've created a library (GetLib) to perform this search as copying it between Crypto and MetaXploit would add a lot of overhead.