Update GetLib.src
Decided that any program that expects to be called remotely should handle including libraries itself. This is because I've realized that the can be multiple instances with different versions of Metaxploit.so on a single system so search for a particular one needs to be handled on a case-by-case basis.